requirements.txt
aiohttp==2.1.0
async-timeout==1.2.1
chardet==3.0.4
inotify==0.2.8
multidict==2.1.6
systemd-python==234
telepot==12.0
urllib3==1.21.1
yarl==0.10.2
